+++ src/resources/views/reports/orders.blade.php
@@ -38,11 +38,11 @@
                     @endforeach
                 </td>
                 <td>
-                    @foreach($order->products as $product)
-                        {{ str_replace(',', '.', $product->name) }} ({{ str_replace(',', '.', $product->sku) }}), 
+                    @foreach ($order->products()->groupBy('name')->get() as $product)
+                        {{ str_replace(',', '.', $product->name) }} ({{ str_replace(',', '.', $product->sku) }}) Qty: {{ $order->products()->where('name', $product->name)->count() }}, 
                     @endforeach
-                    @foreach($order->bundles as $bundle)
-                        {{ str_replace(',', '.', $bundle->name) }} ({{ str_replace(',', '.', $bundle->sku) }}), 
+                    @foreach ($order->bundles()->groupBy('name')->get() as $bundle)
+                        {{ str_replace(',', '.', $bundle->name) }} ({{ str_replace(',', '.', $bundle->sku) }}) Qty: {{ $order->bundles()->where('name', $bundle->name)->count() }}, 
                     @endforeach
                     @foreach ($order->pricelists as $pricelist)
                         {{ str_replace(',', '.', $pricelist->module->name) }}/{{ str_replace(',', '.', $pricelist->membership->name) }} ({{ str_replace(',', '.', $pricelist->module->sku) }}),